Rework clamp interrupt enable code
The code was not prepared for errors: In case pAtmel->readIntMask failed for
whatever reason we wrote an uninitialized variable into controller's interrupt
mask. To fix let's keep interrupt mask and trust controller: If that reports it
has written the mask, we can rely on that. Otherwise if controller reports an
error, spawn a message in journal.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>